services/enrichment/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6
requests==2.33.0
google-generativeai==0.8.6
python-dotenv==1.2.2
